[#6] Rename: 이미지 파일 폴더 추가
