+++ src/routes/home.tsx
@@ -1,6 +1,6 @@
 import styled from "styled-components"
-import logoimg from "../assets/LOGO.svg"
-import introimg from "../assets/INTRO.svg"
+import logoimg from "../assets/home/LOGO.svg"
+import introimg from "../assets/home/INTRO.svg"
 import KakaoBtn from "../components/Kakao"
 
 const Wrapper = styled.div`
